+++ paper/paper.md
@@ -27,6 +27,8 @@ The core problem is that standard persistence mechanisms (pickle/joblib) are **v
 
 The migration proceeds in two stages. **Stage 1 (parity):** the library captures a minimal set of prediction-critical attributes (constructor parameters and other parity-relevant settings) that guarantee parity of outputs across versions; these are used to reconstruct an equivalent estimator in the target version and validated under a strict tolerance (e.g., `max |y_in - y_out| < 1e-2`). **Stage 2 (compatibility):** remaining attributes are serialized with a version-aware policy that gracefully handles additions, removals, and renames so deserialization does not break across releases. This strategy is designed around the practical reality that estimator internals, defaults, and attribute names shift over time—even when the public API remains stable [@sklearn_persistence; @parida2025exportformats].
 
+Concretely, consider version `0.21.3` exposing `param_1, param_2, param_3, param_4` and version `1.7.0` exposing `param_1, param_2, param_3, param_5`. We use `param_1` and `param_2` to enforce identical predictions across versions. Since `param_3` exists in both versions, it is stored in the payload and reassigned on load. For version-specific attributes, the serializer records values for both `param_4` and `param_5`: in `0.21.3`, the payload stores the real value of `param_4` and the **default** value that `param_5` would take in newer versions; in `1.7.0`, it stores the real value of `param_5` and the default value that `param_4` would take in older versions. During deserialization, attributes are assigned using a simple `try/except` (or `hasattr`) so only valid attributes for the target version are set, and missing ones are safely skipped. This design makes migrations resilient to evolutionary changes such as renamed fields (e.g., `affinity → metric`), reorganized tree/boosting internals, and added default parameters across releases.
+
 This submission supports **21 models** across supervised and unsupervised learning:
 
 - **Classification (7):** `DecisionTreeClassifier`, `RandomForestClassifier`, `GradientBoostingClassifier`, `LogisticRegression`, `KNeighborsClassifier`, `SVC`, `MLPClassifier`
@@ -52,9 +54,7 @@ The library targets practitioners and MLOps teams who must migrate or reproduce 
 
 Model persistence and portability are longstanding challenges in applied machine learning. In the scikit-learn ecosystem, the officially recommended mechanisms for saving trained estimators—`pickle` and `joblib`—are explicitly documented as *not* guaranteeing forward or backward compatibility across library versions [@sklearn_persistence]. As a result, serialized models are tightly coupled to the exact scikit-learn release and Python environment in which they were created.
 
-Several alternative approaches partially address related concerns. Interoperability frameworks such as ONNX and PMML enable model exchange across runtimes and languages, but they support only a subset of scikit-learn estimators and often sacrifice access to native APIs, custom preprocessing logic, or exact numerical parity [@parida2025exportformats]. Re-training models after upgrades is a common operational workaround, but it may be infeasible due to missing data, regulatory constraints, or computational cost—particularly in long-lived production systems.
-
-Despite the prevalence of scikit-learn in both research and industry, there is currently no native, Python-centric solution that enables **deterministic migration of fitted scikit-learn estimators across versions while preserving inference behavior**. Existing tools focus either on environment pinning or on cross-framework export, leaving a gap for version-aware, estimator-level portability within the scikit-learn ecosystem itself. `sklearn-migrator` fills this gap by providing a lightweight, inspectable, and version-resilient mechanism for migrating models across scikit-learn releases without retraining.
[…]
