### 2026-03-27 #9662 @crvernon
:wave: @jbytecode are you able to take this one on as editor?
