### 2026-03-27 #9662 @jbytecode
@crvernon - sure!
