### 2026-03-27 #9662 @jbytecode
@anvaldes - Hi, I'm the handling editor of your submission. Thank you for submitting your work to JOSS.

Your manuscript markdown file must be end with a `# References` line to make a title to the bibliography. Please check and fix.

Initially, I'll try to invite and wait a response from a list of candidate reviewers that our editorial bot suggests. Meanwhile, you can prepare a list of potential reviewers. In case we fail to assign two reviewers using the suggestions, we can use your list.
