### 2026-04-03 #9662 @anvaldes
Hi @jbytecode,

Thanks for pointing this out. I’ve added the missing `# References` section at the end of the manuscript and regenerated the PDF.

Please let me know if anything else needs adjustment.

Best regards,
Alberto
