### 2026-04-03 #9662 @anvaldes
Hi @jbytecode,

Thanks for the update. I would like to suggest a few potential reviewers who might be a good fit for this work:

* reidjohnson (quantile-forest; tree-based methods and sklearn extensions)
* jncraton (reviewer in sklearn-related JOSS papers)
* papachristoumarios (reviewer in ML tooling libraries)

Please let me know if I should suggest additional candidates.

Best regards,
Alberto
