### 2026-04-03 #10374 @editorialbot
[REVIEW]: sklearn-migrator: Cross-version migration of scikit-learn models for reproducible MLOps
**Submitting author:** @anvaldes (<a href="http://orcid.org/0009-0000-0752-8519">Alberto Andres Valdes Gonzalez</a>)
**Repository:** https://github.com/anvaldes/sklearn-migrator
**Branch with paper.md** (empty if default branch): main
**Version:** v0.21.1
