### 2026-04-04 #10374 @smk508
I left comments in the form of issues. My main feedback regards testing. I think the reason this problem is so difficult and largely unaddressed is exactly because of how much variation there can be between different sklearn model types and hence the number of different tests you would have to do to prove that a migrator such as this works well. Because of that, demonstrating correct migrations across a range of version and model pairs would be the most important piece of this contribution.  At the very least, the claims in the README about 1000+ migration pairs should be verifiable through CICD and I could not find any test scripts to do this. I also have questions about how deeply individual migration functions need to be tested in order to cover all of the minor features and hyperparameters within a specific model family.
