+++ integration/env_in/requirements.txt
@@ -1,4 +1,0 @@
-scikit-learn==1.0.0
-numpy==1.19.5
-pandas==1.4.4
-joblib==1.5.1
