+++ .github/workflows/tests.yml
@@ -1,5 +1,4 @@
 name: tests
-
 on:
   push:
     branches: ["**"]
@@ -13,7 +12,6 @@ jobs:
       fail-fast: false
       matrix:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
-
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -46,44 +44,3 @@ jobs:
             pytest-junit.xml
             coverage.xml
           if-no-files-found: ignore
-
-  integration-tests:
-    runs-on: ubuntu-latest
-    needs: [test]
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v3
-
-      - name: Set up Python (host for compare_preds.py)
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.12"
-
-      - name: Install minimal deps (host)
-        run: |
-          python -m pip install --upgrade pip
-          pip install numpy
-
-      - name: Build integration images
-        run: |
-          docker build -t sklearn-migrator-env-in  -f integration/env_in/Dockerfile .
-          docker build -t sklearn-migrator-env-out -f integration/env_out/Dockerfile .
-
-      - name: Run integration script
-        run: |
-          set -euo pipefail
-          chmod +x integration/run.sh
-          ./integration/run.sh
-
-      - name: Upload integration artifacts
-        if: always()
-        uses: actions/upload-artifact@v4
-        with:
-          name: integration-artifacts
-          path: integration/_artifacts
-          if-no-files-found: ignore
-
