+++ .github/workflows/tests.yml
@@ -44,3 +44,25 @@ jobs:
             pytest-junit.xml
             coverage.xml
           if-no-files-found: ignore
+
+  build-input-images:
+    runs-on: ubuntu-latest
+    needs: [test]
+    strategy:
+      fail-fast: false
+      matrix:
+        version: ["0.21.3"]
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Login to GHCR
+        run: echo "${{ secrets.GITHUB_TOKEN }}" | docker login ghcr.io -u ${{ github.actor }} --password-stdin
+
+      - name: Build and push input image
+        run: |
+          docker build \
+            -f integration/environments/input/${{ matrix.version }}/Dockerfile_input \
+            -t ghcr.io/${{ github.repository }}/sklearn-migrator-input-${{ matrix.version }}:latest \
+            .
+          docker push ghcr.io/${{ github.repository }}/sklearn-migrator-input-${{ matrix.version }}:latest
