+++ integration/environments/input/1.1.3/requirements_input.txt
@@ -1,0 +1,4 @@
+numpy==1.19.5
+scikit-learn==1.1.3
+pandas==1.4.4
+joblib==1.5.1
