+++ integration/environments/input/1.7.2/requirements_input.txt
@@ -1,0 +1,3 @@
+scikit-learn==1.7.2
+pandas==2.3.0
+joblib==1.5.1
