+++ integration/environments/output/1.1.0/requirements_output.txt
@@ -1,0 +1,4 @@
+numpy==1.19.5
+scikit-learn==1.1.0
+pandas==1.4.4
+joblib==1.5.1
