+++ integration/environments/output/1.3.2/requirements_output.txt
@@ -1,0 +1,4 @@
+scikit-learn==1.3.2
+numpy==1.26.4
+pandas==2.3.0
+joblib==1.5.1
