+++ .github/workflows/tests.yml
@@ -102,3 +102,62 @@ jobs:
       - name: Push input image
         run: |
           docker push ghcr.io/${{ github.repository }}/sklearn-migrator-input-${{ matrix.version }}:latest
+  
+  build-output-images:
+    runs-on: ubuntu-latest
+    needs: [build-input-images]
+    strategy:
+      fail-fast: false
+      matrix:
+        version: [
+          "0.21.3",
+          "0.22.0",
+          "0.22.1",
+          "0.23.0",
+          "0.23.1",
+          "0.23.2",
+          "0.24.0",
+          "0.24.1",
+          "0.24.2",
+          "1.0.0",
+          "1.0.1",
+          "1.0.2",
+          "1.1.0",
+          "1.1.1",
+          "1.1.2",
+          "1.1.3",
+          "1.2.0",
+          "1.2.1",
+          "1.2.2",
+          "1.3.0",
+          "1.3.1",
+          "1.3.2",
+          "1.4.0",
+          "1.4.2",
+          "1.5.0",
+          "1.5.1",
+          "1.5.2",
+          "1.6.0",
+          "1.6.1",
+          "1.7.0",
+          "1.7.1",
+          "1.7.2"
+        ]
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Login to GHCR
+        run: echo "${{ secrets.GITHUB_TOKEN }}" | docker login ghcr.io -u ${{ github.actor }} --password-stdin
+
+      - name: Build output image
+        run: |
+          docker build \
+            --build-arg VERSION=${{ matrix.version }} \
+            -f integration/environments/output/${{ matrix.version }}/Dockerfile_output \
+            -t ghcr.io/${{ github.repository }}/sklearn-migrator-output-${{ matrix.version }}:latest \
+            .
+
+      - name: Push output image
+        run: |
+          docker push ghcr.io/${{ github.repository }}/sklearn-migrator-output-${{ matrix.version }}:latest
