+++ integration/scripts/run_svm_reg.py
@@ -1,0 +1,173 @@
+import os
+import io
+import sys
+import time
+import shutil
+import subprocess
+import pandas as pd
+
+# -----------------------------------------------------------------------
+# Config
+# -----------------------------------------------------------------------
+name_model = "svm_reg"
+workspace  = os.environ["GITHUB_WORKSPACE"]
+repository = os.environ["GITHUB_REPOSITORY"]
+
+base_input_dir  = os.path.join(workspace, f"integration/input_{name_model}")
+base_output_dir = os.path.join(workspace, f"integration/output_{name_model}")
+
+input_versions = [
+    "0.21.3", "0.22.0"
+]
+
+output_versions = [
+    "1.7.1",  "1.7.2"
+]
+
+REGISTRY = f"ghcr.io/{repository}"
+
+# -----------------------------------------------------------------------
+# Tee
+# -----------------------------------------------------------------------
+class Tee:
+    def __init__(self, file):
+        self.file   = file
+        self.stdout = sys.stdout
+
+    def write(self, data):
+        self.stdout.write(data)
+        self.file.write(data)
+
+    def flush(self):
+        self.stdout.flush()
+        self.file.flush()
+
+# -----------------------------------------------------------------------
+# container_execution
+# -----------------------------------------------------------------------
+def container_execution(inp_ver, out_ver):
+    print(f"INPUT: {inp_ver}")
+    print(f"OUTPUT: {out_ver}")
+
+    name_input_image  = f"{REGISTRY}/sklearn-migrator-input-{inp_ver}:latest"
+    name_output_image = f"{REGISTRY}/sklearn-migrator-output-{out_ver}:latest"
+
+    shutil.rmtree(base_input_dir,  ignore_errors=True)
+    shutil.rmtree(base_output_dir, ignore_errors=True)
+    os.makedirs(base_input_dir,  exist_ok=True, mode=0o777)
+    os.makedirs(base_output_dir, exist_ok=True, mode=0o777)
+
+    # 1. Run INPUT container
+    try:
+        result_input = subprocess.run([
+            "docker", "run", "--rm",
+            "-v", f"{base_input_dir}:/input",
+            name_input_image,
+            f"input_{name_model}.py"
+        ], check=True, capture_output=True, text=True)
+        print(result_input.stdout)
+        print(result_input.stderr)
+        print("INPUT execution: ✅")
+    except subprocess.CalledProcessError as e:
+        print("INPUT execution: ❌")
+        print(f"STDOUT: {e.stdout}")
+        print(f"STDERR: {e.stderr}")
+        return None
+
+    # 2. Copy serialized_model
+    try:
+        src = os.path.join(base_input_dir,  "serialized_model.json")
+        dst = os.path.join(base_output_dir, "serialized_model.json")
+        os.makedirs(os.path.dirname(dst), exist_ok=True)
+        shutil.copyfile(src, dst)
+    except Exception as e:
+        print(f"COPY serialized_model: ❌\nERROR: {e}")
+        return None
+
+    # 3. Run OUTPUT container
+    try:
+        result_output = subprocess.run([
+            "docker", "run", "--rm",
+            "-v", f"{base_output_dir}:/output",
+            name_output_image,
+            f"output_{name_model}.py"
+        ], check=True, capture_output=True, text=True)
+        print(result_output.stdout)
+        print(result_output.stderr)
+        print("OUTPUT execution: ✅")
+    except subprocess.CalledProcessError as e:
+        print("OUTPUT execution: ❌")
+        print(f"STDOUT: {e.stdout}")
+        print(f"STDERR: {e.stderr}")
+        return None
+
+    # 4. Comparison
+    try:
+        df_input  = pd.read_csv(f"{base_input_dir}/y_pred_input.csv")
[…]
