+++ .github/workflows/tests.yml
@@ -161,3 +161,64 @@ jobs:
       - name: Push output image
         run: |
           docker push ghcr.io/${{ github.repository }}/sklearn-migrator-output-${{ matrix.version }}:latest
+  
+  integration-tests:
+    runs-on: ubuntu-latest
+    needs: [build-input-images, build-output-images]
+    strategy:
+      fail-fast: false
+      matrix:
+        name_model: ["svm_reg"]
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Login to GHCR
+        run: echo "${{ secrets.GITHUB_TOKEN }}" | docker login ghcr.io -u ${{ github.actor }} --password-stdin
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install pandas
+
+      - name: Pull all input images
+        run: |
+          for version in 0.21.3 0.22.0 0.22.1 0.23.0 0.23.1 0.23.2 0.24.0 0.24.1 0.24.2 \
+                         1.0.0 1.0.1 1.0.2 1.1.0 1.1.1 1.1.2 1.1.3 1.2.0 1.2.1 1.2.2 \
+                         1.3.0 1.3.1 1.3.2 1.4.0 1.4.2 1.5.0 1.5.1 1.5.2 1.6.0 1.6.1 \
+                         1.7.0 1.7.1 1.7.2; do
+            docker pull ghcr.io/${{ github.repository }}/sklearn-migrator-input-${version}:latest
+          done
+
+      - name: Pull all output images
+        run: |
+          for version in 0.21.3 0.22.0 0.22.1 0.23.0 0.23.1 0.23.2 0.24.0 0.24.1 0.24.2 \
+                         1.0.0 1.0.1 1.0.2 1.1.0 1.1.1 1.1.2 1.1.3 1.2.0 1.2.1 1.2.2 \
+                         1.3.0 1.3.1 1.3.2 1.4.0 1.4.2 1.5.0 1.5.1 1.5.2 1.6.0 1.6.1 \
+                         1.7.0 1.7.1 1.7.2; do
+            docker pull ghcr.io/${{ github.repository }}/sklearn-migrator-output-${version}:latest
+          done
+
+      - name: Run integration test
+        run: python integration/scripts/run_${{ matrix.name_model }}.py
+
+      - name: Upload logs
+        if: always()
+        uses: actions/upload-artifact@v4
+        with:
+          name: logs-${{ matrix.name_model }}
+          path: logs/
+          if-no-files-found: ignore
+
+      - name: Upload performance
+        if: always()
+        uses: actions/upload-artifact@v4
+        with:
+          name: performance-${{ matrix.name_model }}
+          path: performance/performance_${{ matrix.name_model }}.csv
+          if-no-files-found: ignore
