+++ .github/workflows/tests.yml
@@ -7,6 +7,7 @@ on:
 
 jobs:
   test:
+    if: false
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
@@ -46,6 +47,7 @@ jobs:
           if-no-files-found: ignore
 
   build-input-images:
+    if: false
     runs-on: ubuntu-latest
     needs: [test]
     strategy:
@@ -104,6 +106,7 @@ jobs:
           docker push ghcr.io/${{ github.repository }}/sklearn-migrator-input-${{ matrix.version }}:latest
   
   build-output-images:
+    if: false
     runs-on: ubuntu-latest
     needs: [build-input-images]
     strategy:
@@ -163,6 +166,7 @@ jobs:
           docker push ghcr.io/${{ github.repository }}/sklearn-migrator-output-${{ matrix.version }}:latest
   
   integration-tests:
+    if: true
     runs-on: ubuntu-latest
     needs: [build-input-images, build-output-images]
     strategy:
@@ -188,19 +192,13 @@ jobs:
 
       - name: Pull all input images
         run: |
-          for version in 0.21.3 0.22.0 0.22.1 0.23.0 0.23.1 0.23.2 0.24.0 0.24.1 0.24.2 \
-                         1.0.0 1.0.1 1.0.2 1.1.0 1.1.1 1.1.2 1.1.3 1.2.0 1.2.1 1.2.2 \
-                         1.3.0 1.3.1 1.3.2 1.4.0 1.4.2 1.5.0 1.5.1 1.5.2 1.6.0 1.6.1 \
-                         1.7.0 1.7.1 1.7.2; do
+          for version in 0.21.3 0.22.0; do
             docker pull ghcr.io/${{ github.repository }}/sklearn-migrator-input-${version}:latest
           done
 
       - name: Pull all output images
         run: |
-          for version in 0.21.3 0.22.0 0.22.1 0.23.0 0.23.1 0.23.2 0.24.0 0.24.1 0.24.2 \
-                         1.0.0 1.0.1 1.0.2 1.1.0 1.1.1 1.1.2 1.1.3 1.2.0 1.2.1 1.2.2 \
-                         1.3.0 1.3.1 1.3.2 1.4.0 1.4.2 1.5.0 1.5.1 1.5.2 1.6.0 1.6.1 \
-                         1.7.0 1.7.1 1.7.2; do
+          for version in 1.7.1 1.7.2; do
             docker pull ghcr.io/${{ github.repository }}/sklearn-migrator-output-${version}:latest
           done
 
@@ -222,3 +220,11 @@ jobs:
           name: performance-${{ matrix.name_model }}
           path: performance/performance_${{ matrix.name_model }}.csv
           if-no-files-found: ignore
+      
+      - name: Commit performance CSV
+        run: |
+          git config user.name "github-actions"
+          git config user.email "github-actions@github.com"
+          git add performance/
+          git commit -m "perf: add performance results for ${{ matrix.name_model }}"
+          git push
