+++ .github/workflows/tests.yml
@@ -7,7 +7,7 @@ on:
 
 jobs:
   test:
-    if: false
+    if: true
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
@@ -47,7 +47,7 @@ jobs:
           if-no-files-found: ignore
 
   build-input-images:
-    if: false
+    if: true
     runs-on: ubuntu-latest
     needs: [test]
     strategy:
@@ -106,7 +106,7 @@ jobs:
           docker push ghcr.io/${{ github.repository }}/sklearn-migrator-input-${{ matrix.version }}:latest
   
   build-output-images:
-    if: false
+    if: true
     runs-on: ubuntu-latest
     needs: [build-input-images]
     strategy:
@@ -168,6 +168,7 @@ jobs:
   integration-tests:
     if: true
     runs-on: ubuntu-latest
+    needs: [build-output-images]
     strategy:
       fail-fast: false
       matrix:
