+++ .github/workflows/tests.yml
@@ -192,13 +192,19 @@ jobs:
 
       - name: Pull all input images
         run: |
-          for version in 0.21.3 0.22.0; do
+          for version in 0.21.3 0.22.0 0.22.1 0.23.0 0.23.1 0.23.2 0.24.0 0.24.1 0.24.2 \
+                         1.0.0 1.0.1 1.0.2 1.1.0 1.1.1 1.1.2 1.1.3 1.2.0 1.2.1 1.2.2 \
+                         1.3.0 1.3.1 1.3.2 1.4.0 1.4.2 1.5.0 1.5.1 1.5.2 1.6.0 1.6.1 \
+                         1.7.0 1.7.1 1.7.2; do
             docker pull ghcr.io/${{ github.repository }}/sklearn-migrator-input-${version}:latest
           done
 
       - name: Pull all output images
         run: |
-          for version in 1.7.1 1.7.2; do
+          for version in 0.21.3 0.22.0 0.22.1 0.23.0 0.23.1 0.23.2 0.24.0 0.24.1 0.24.2 \
+                         1.0.0 1.0.1 1.0.2 1.1.0 1.1.1 1.1.2 1.1.3 1.2.0 1.2.1 1.2.2 \
+                         1.3.0 1.3.1 1.3.2 1.4.0 1.4.2 1.5.0 1.5.1 1.5.2 1.6.0 1.6.1 \
+                         1.7.0 1.7.1 1.7.2; do
             docker pull ghcr.io/${{ github.repository }}/sklearn-migrator-output-${version}:latest
           done
 
