+++ integration/scripts/run_lasso_reg.py
@@ -1,0 +1,235 @@
+import os
+import io
+import sys
+import time
+import shutil
+import subprocess
+import pandas as pd
+
+# -----------------------------------------------------------------------
+# Config
+# -----------------------------------------------------------------------
+name_model = "lasso_reg"
+workspace  = os.environ["GITHUB_WORKSPACE"]
+repository = os.environ["GITHUB_REPOSITORY"]
+
+base_input_dir  = os.path.join(workspace, f"integration/input_{name_model}")
+base_output_dir = os.path.join(workspace, f"integration/output_{name_model}")
+
+input_versions = [
+          "0.21.3",
+          "0.22.0",
+          "0.22.1",
+          "0.23.0",
+          "0.23.1",
+          "0.23.2",
+          "0.24.0",
+          "0.24.1",
+          "0.24.2",
+          "1.0.0",
+          "1.0.1",
+          "1.0.2",
+          "1.1.0",
+          "1.1.1",
+          "1.1.2",
+          "1.1.3",
+          "1.2.0",
+          "1.2.1",
+          "1.2.2",
+          "1.3.0",
+          "1.3.1",
+          "1.3.2",
+          "1.4.0",
+          "1.4.2",
+          "1.5.0",
+          "1.5.1",
+          "1.5.2",
+          "1.6.0",
+          "1.6.1",
+          "1.7.0",
+          "1.7.1",
+          "1.7.2"
+        ]
+
+output_versions = [
+          "0.21.3",
+          "0.22.0",
+          "0.22.1",
+          "0.23.0",
+          "0.23.1",
+          "0.23.2",
+          "0.24.0",
+          "0.24.1",
+          "0.24.2",
+          "1.0.0",
+          "1.0.1",
+          "1.0.2",
+          "1.1.0",
+          "1.1.1",
+          "1.1.2",
+          "1.1.3",
+          "1.2.0",
+          "1.2.1",
+          "1.2.2",
+          "1.3.0",
+          "1.3.1",
+          "1.3.2",
+          "1.4.0",
+          "1.4.2",
+          "1.5.0",
+          "1.5.1",
+          "1.5.2",
+          "1.6.0",
+          "1.6.1",
+          "1.7.0",
+          "1.7.1",
+          "1.7.2"
+        ]
+
+REGISTRY = f"ghcr.io/{repository}"
+
+# -----------------------------------------------------------------------
+# Tee
+# -----------------------------------------------------------------------
+class Tee:
+    def __init__(self, file):
+        self.file   = file
+        self.stdout = sys.stdout
+
+    def write(self, data):
+        self.stdout.write(data)
+        self.file.write(data)
+
+    def flush(self):
+        self.stdout.flush()
+        self.file.flush()
+
+# -----------------------------------------------------------------------
+# container_execution
+# -----------------------------------------------------------------------
+def container_execution(inp_ver, out_ver):
+    print(f"INPUT: {inp_ver}")
+    print(f"OUTPUT: {out_ver}")
+
+    name_input_image  = f"{REGISTRY}/sklearn-migrator-input-{inp_ver}:latest"
+    name_output_image = f"{REGISTRY}/sklearn-migrator-output-{out_ver}:latest"
+
+    shutil.rmtree(base_input_dir,  ignore_errors=True)
+    shutil.rmtree(base_output_dir, ignore_errors=True)
+    os.makedirs(base_input_dir,  exist_ok=True, mode=0o777)
+    os.makedirs(base_output_dir, exist_ok=True, mode=0o777)
+
+    # 1. Run INPUT container
+    try:
+        result_input = subprocess.run([
+            "docker", "run", "--rm",
+            "-v", f"{base_input_dir}:/input",
+            name_input_image,
[…]
