### 2026-04-13 #10374 @anvaldes
Hi @smk508, thanks again for your detailed feedback.

I wanted to share a quick update to let you know that I am actively working on the points you raised, particularly around testing and validation.

Over the past days, I have been focusing on building integration tests to systematically compare model behavior across different scikit-learn versions.

I will continue addressing the issues you opened, and I’ll provide more detailed updates as I make progress on each of them.

Thanks again for the valuable feedback — it’s been very helpful in improving the quality of the project.
