+++ sklearn_migrator/dimension/pca.py
@@ -1,3 +1,4 @@
+import warnings
 import numpy as np
 from sklearn.decomposition import PCA
 
@@ -84,13 +85,19 @@ def deserialize_pca(data, version_out):
 
     for af in all_features:
         if af not in other_params:
-            continue
-
+            continue  # field not present in this sklearn version
         value = other_params[af]
-
         if af in array_fields and value is not None and not isinstance(value, np.ndarray):
             value = np.array(value)
-
-        new_model.__dict__[af] = value
+        try:
+            new_model.__dict__[af] = value
+        except AttributeError:
+            pass  # attribute not settable in this sklearn version
+        except Exception as e:
+            warnings.warn(
+                f"Could not set field '{af}' on {type(new_model).__name__}: "
+                f"{type(e).__name__}: {e}. Field will be skipped.",
+                UserWarning,
+            )
 
     return new_model
