+++ sklearn_migrator/regression/random_forest_reg.py
@@ -1,3 +1,4 @@
+import warnings
 import numpy as np
 from sklearn.ensemble import RandomForestRegressor
 from .decision_tree_reg import serialize_decision_tree_reg
@@ -97,8 +98,16 @@ def deserialize_random_forest_reg(data, version_out):
     for af in all_features:
         try:
             new_model.__dict__[af] = data['other_params'][af]
-        except:
-            pass
+        except KeyError:
+            pass  # field not present in this sklearn version
+        except AttributeError:
+            pass  # attribute not settable in this sklearn version
+        except Exception as e:
+            warnings.warn(
+                f"Could not set field '{af}' on {type(new_model).__name__}: "
+                f"{type(e).__name__}: {e}. Field will be skipped.",
+                UserWarning,
+            )
 
     n_features = (data['other_params']['n_features'] or data['other_params']['n_features_in'])
 
