+++ sklearn_migrator/classification/random_forest_clf.py
@@ -1,3 +1,4 @@
+import warnings
 import numpy as np
 from sklearn.ensemble import RandomForestClassifier
 from .decision_tree_clf import serialize_decision_tree_clf
@@ -99,8 +100,16 @@ def deserialize_random_forest_clf(data, version_out):
     for af in all_features:
         try:
             new_model.__dict__[af] = data['other_params'][af]
-        except:
-            pass
+        except KeyError:
+            pass  # field not present in this sklearn version
+        except AttributeError:
+            pass  # attribute not settable in this sklearn version
+        except Exception as e:
+            warnings.warn(
+                f"Could not set field '{af}' on {type(new_model).__name__}: "
+                f"{type(e).__name__}: {e}. Field will be skipped.",
+                UserWarning,
+            )
 
     n_features = (data['other_params']['n_features'] or data['other_params']['n_features_in'])
 
