### 2026-04-18 f92c18b
Adding docstring and annotations
