+++ sklearn_migrator/regression/decision_tree_reg.py
@@ -26,29 +26,74 @@ all_features = [
 ]
 
 
-def version_tuple(version):
+def version_tuple(version: str) -> tuple:
+    """
+    Convert a version string into a comparable tuple of integers.
 
-  version_split = version.split('.')
+    Parameters
+    ----------
+    version : str
+        Version string (e.g. '1.2.0').
 
-  if len(version_split) == 1:
-    new_version = (int(version_split[0]), 0, 0)
-  elif len(version_split) == 2:
-    new_version = (int(version_split[0]), int(version_split[1]), 0)
-  elif len(version_split) == 3:
-    new_version = (int(version_split[0]), int(version_split[1]), int(version_split[2]))
-  else:
-    new_version = 'Formato no valido'
+    Returns
+    -------
+    tuple
+        Tuple of integers (major, minor, patch).
+    """
 
-  return new_version
+    version_split = version.split('.')
+
+    if len(version_split) == 1:
+        new_version = (int(version_split[0]), 0, 0)
+    elif len(version_split) == 2:
+        new_version = (int(version_split[0]), int(version_split[1]), 0)
+    elif len(version_split) == 3:
+        new_version = (int(version_split[0]), int(version_split[1]), int(version_split[2]))
+    else:
+        new_version = 'Formato no valido'
+
+    return new_version
 
 
-def _get_extended_nodes(nodes, version_in):
+def _get_extended_nodes(nodes: list, version_in: str) -> list:
+    """
+    Extend node tuples with a placeholder field for sklearn versions < 1.3.
+
+    Parameters
+    ----------
+    nodes : list
+        List of node tuples from the tree state.
+    version_in : str
+        The sklearn version used to train the model.
+
+    Returns
+    -------
+    list
+        List of node tuples, extended if necessary.
+    """
+
     if (version_tuple('0.21.3') <= version_tuple(version_in)) and (version_tuple(version_in) < version_tuple('1.3')):
         return [node + (0,) for node in nodes]
     return nodes
 
 
-def _build_dtype_dict(dtypes, version_in):
+def _build_dtype_dict(dtypes: np.dtype, version_in: str) -> dict:
+    """
+    Build a dictionary describing the node dtype structure of the tree.
+
+    Parameters
+    ----------
+    dtypes : np.dtype
+        The dtype of the nodes array from the tree state.
+    version_in : str
+        The sklearn version used to train the model.
+
+    Returns
+    -------
+    dict
+        Dictionary with field names, formats, offsets and itemsize.
+    """
+
     field_names = dtypes.names
     formats = [dtypes.fields[name][0] for name in field_names]
     offsets = [dtypes.fields[name][1] for name in field_names]
@@ -70,7 +115,22 @@ def _build_dtype_dict(dtypes, version_in):
     }
 
 
-def _get_metadata(model, version_in):
+def _get_metadata(model: DecisionTreeRegressor, version_in: str) -> dict:
+    """
+    Extract feature metadata from a fitted DecisionTreeRegressor.
+
+    Parameters
+    ----------
+    model : DecisionTreeRegressor
+        A fitted scikit-learn DecisionTreeRegressor instance.
+    version_in : str
+        The sklearn version used to train the model.
+
+    Returns
+    -------
+    dict
+        Dictionary with n_features_in, n_features, n_classes and n_outputs.
+    """
 
     dict_metadata = {}
 
@@ -94,7 +154,23 @@ def _get_metadata(model, version_in):
     return dict_metadata
 
 
-def serialize_decision_tree_reg(model, version_in):
+def serialize_decision_tree_reg(model: DecisionTreeRegressor, version_in: str) -> dict:
+    """
[…]
