+++ paper/paper.md
@@ -50,9 +50,19 @@ The library targets practitioners and MLOps teams who must migrate or reproduce 
 
 # State of the field
 
-Model persistence and portability are longstanding challenges in applied machine learning. In the scikit-learn ecosystem, the officially recommended mechanisms for saving trained estimators—`pickle` and `joblib`—are explicitly documented as *not* guaranteeing forward or backward compatibility across library versions [@sklearn_persistence]. As a result, serialized models are tightly coupled to the exact scikit-learn release and Python environment in which they were created.
+Model persistence and portability are longstanding challenges in applied machine learning, with documented community demand for cross-version compatibility solutions [@sklearn_issue10319]. In the scikit-learn ecosystem, the officially recommended mechanisms for saving trained estimators—`pickle` and `joblib`—are explicitly documented as *not* guaranteeing forward or backward compatibility across library versions [@sklearn_persistence]. As a result, serialized models are tightly coupled to the exact scikit-learn release and Python environment in which they were created.
 
-Several alternative approaches partially address related concerns. Interoperability frameworks such as ONNX and PMML enable model exchange across runtimes and languages, but they support only a subset of scikit-learn estimators and often sacrifice access to native APIs, custom preprocessing logic, or numerical parity [@parida2025exportformats]. Re-training models after upgrades is a common workaround, but it may be infeasible due to missing data, regulatory constraints, or computational cost.
+Several alternative approaches partially address related concerns, but each comes with significant limitations in practice:
+
+- **pickle/joblib**: the default persistence mechanism, but explicitly version-fragile. A model saved under one release may fail to load or behave differently under another [@sklearn_persistence].
+- **skops**: designed for secure model sharing and audit trails using a safer serialization format than pickle, but does not address cross-version compatibility — a model serialized with skops under one sklearn version may still fail to load under another.
+- **sklearn-onnx**: converts scikit-learn models to the ONNX format for cross-runtime interoperability, but supports only a subset of estimators, does not preserve the native scikit-learn API, and introduces a dependency on the ONNX runtime [@parida2025exportformats].
+- **PMML**: a standard interchange format with broad tool support, but with limited scikit-learn estimator coverage and no guarantee of numerical parity.
+- **Re-training**: a common workaround, but often infeasible due to missing data, regulatory constraints, or computational cost.
+
+Beyond technical limitations, there is an important security dimension to this problem. Pinning old scikit-learn versions to avoid the cost of retraining or migration is not a cost-free decision: it accumulates CVEs and security debt over time, exposing production systems to known vulnerabilities. This makes the case for a migration tool particularly compelling in organizations where dependency upgrades are driven by security patching cycles rather than feature adoption.
+
+`sklearn-migrator` addresses this gap directly by enabling native, Python-centric cross-version migration without requiring re-training, runtime conversion, or version pinning — while maintaining full compatibility with the scikit-learn API.
 
 # Research Impact Statement
 
@@ -94,9 +104,11 @@ To keep payloads portable, the library restricts values to JSON-encodable primit
 
 - **Environment isolation** (e.g., containers) to ensure `version_in` and `version_out` represent real installations.
 - **Fixed synthetic datasets** for deterministic evaluation.
-- **Strict parity checks** comparing source and migrated predictions under a tolerance (e.g., `1e-4`).
+- **Strict parity checks** comparing source and migrated predictions under a tolerance (e.g., `1e-2`).
 
[…]
