### 2026-04-19 9e21dc4
Paper modifications
