### 2026-04-19 019e3c8
Returning to SVC implementation
