### 2026-04-20 7e77b46
Paper modifications
