### 2026-04-22 #10374 @anvaldes
Hi @smk508 and @jbytecode,
I've addressed all the issues raised during the review. Here's a summary of what was done:

#107 – Unified all tolerance references in paper.md to 1e-2, consistent with the CI/CD pipeline.
#108 – Fixed the formatting error in the limitations section so the bullet points are correctly separated.
#109 – Expanded the "State of the field" section with explicit comparisons to pickle/joblib, skops, sklearn-onnx, PMML, and re-training, including a reference to scikit-learn/scikit-learn#10319 and a discussion of the CVE accumulation risk from dependency pinning.
#110 – Added NumPy-style docstrings and type annotations to all function definitions across the codebase.
#111 – Replaced all bare except: pass blocks in deserializer loops with explicit handling: KeyError and AttributeError are silently skipped, while any other exception emits a warnings.warn().
#112 – The CI/CD pipeline now runs integration tests for all 21 supported models across all 32×32 = 1,024 version pairs (21,504 executions total), with results uploaded as artifacts on every push.
#113 – Clarified in both the documentation and paper.md that each model is validated under a representative configuration, and that exhaustive combinatorial testing is outside the current scope.
