### 2026-04-27 #10374 @smk508
thanks @anvaldes I left some follow up comments on some of the items please take a look
