### 2026-04-28 8f06b8e
Changes on tolerance and try except logic
