+++ tests/regression/test_mlp_reg.py
@@ -27,7 +27,7 @@ def test_mlp_reg():
     y_pred = model.predict(X)
     y_pred_new = new_model.predict(X)
 
-    threshold = 0.001
+    threshold = 1e-2
 
     assert (abs(y_pred - y_pred_new).max() <= threshold)
 
