### 2026-04-28 #10374 @anvaldes
Hi @smk508  — all follow-up comments have been addressed:

- Issue 107 = (inconsistent tolerances): `CONTRIBUTING.md` and unit tests updated to `1e-2` throughout.
- Issue 111 = (broad bare `except` blocks): remaining `except: pass` blocks updated across all deserializers.
- Issue 112 =  (property-based tests claim): removed from paper and documentation.

Thanks again for the careful review.
