### 2026-05-05 #10374 @anvaldes
Hi @solegalli and @jbytecode,I have addressed both suggestions from @solegalli:

1. README example fixed — The serialization example now includes self-contained data generation using make_regression and train_test_split, and the deserialization example reads X_test from a saved CSV file, making both code blocks fully copy-pasteable and runnable out of the box.

2. Dedicated API documentation — A new API.md file has been added to the repository, documenting all 42 public functions (21 serialize + 21 deserialize) across all supported models. Each entry includes the function signature, parameter types and descriptions, return type, and a usage example.
